+++ src/renderer/utils/musicSdk/kg/index.js
@@ -6,10 +6,10 @@ import pic from './pic'
 import lyric from './lyric'
 import hotSearch from './hotSearch'
 import comment from './comment'
-import tipSearch from './tipSearch'
+// import tipSearch from './tipSearch'
 
 const kg = {
-  tipSearch,
+  // tipSearch,
   leaderboard,
   songList,
   musicSearch,
