+++ src/renderer/utils/musicSdk/utils.js
@@ -22,7 +22,6 @@ export const getHostIp = hostname => {
   })
 }
 
-
 export const dnsLookup = (hostname, options, callback) => {
   const result = getHostIp(hostname)
   if (result) return callback(null, result.address, result.family)
@@ -34,24 +33,24 @@ export const dnsLookup = (hostname, options, callback) => {
 /**
  * 格式化歌手
  * @param singers 歌手数组
- * @param obj 读取的数据
- * @param join 插入的字符
+ * @param nameKey 歌手名键值
+ * @param join 歌手分割字符
  */
-export const formatSingerName = (singers, obj = 'name', join = '、') => {
-  if (typeof singers === 'string') {
+export const formatSingerName = (singers, nameKey = 'name', join = '、') => {
+  if (typeof singers == 'string') {
     try {
       singers = JSON.parse(singers)
     } catch (err) {
       return decodeName(singers)
     }
+  } else if (Array.isArray(singers)) {
+    const singer = []
+    singers.forEach(item => {
+      let name = item[nameKey]
+      if (!name) return
+      singer.push(name)
+    })
+    return decodeName(singer.join(join))
   }
-  if (!Array.isArray(singers)) return decodeName(singers)
-
-  const singer = []
-  singers.forEach(item => {
-    let name = item[obj]
-    if (!name) return
-    singer.push(name)
-  })
-  return decodeName(singer.join(join))
+  return decodeName(String(singers ?? ''))
 }
