+++ src/renderer/utils/musicSdk/wy/index.js
@@ -6,10 +6,10 @@ import musicSearch from './musicSearch'
 import songList from './songList'
 import hotSearch from './hotSearch'
 import comment from './comment'
-import tipSearch from './tipSearch'
+// import tipSearch from './tipSearch'
 
 const wy = {
-  tipSearch,
+  // tipSearch,
   leaderboard,
   musicSearch,
   songList,
